goldfix/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.43
joblib==1.4.2
scikit-learn==1.5.1
tensorflow==2.17.0
python-dotenv==1.0.1
